Bug 1035786 - Avoid warning-as-error sandbox build failure with an explicit cast. r=gdestuynder
getpid() is never negative, so this is safe.

UltraBlame original commit: 78ebcfff12347b6cd921ddb8d240a89912b106d3

+++ security/sandbox/linux/SandboxFilter.cpp
@@ -210,7 +210,7 @@ SandboxFilterImpl::Build() {
 
   
   
-  Allow(SYSCALL_WITH_ARG(tgkill, 0, getpid()));
+  Allow(SYSCALL_WITH_ARG(tgkill, 0, uint32_t(getpid())));
 
   
 #ifdef MOZ_WIDGET_GONK
